自动落锁相关配置移入SETTING模块
