+++ user/inc/data.h
@@ -10,20 +10,5 @@
 
 #include <eat_type.h>
 
-typedef struct
-{
-    eat_bool isAutodefendFixed;
-    unsigned char AutodefendPeriod;
-}LOCAL_DATA;
-
-
-
-
-void set_autodefend_state(eat_bool fixed);
-void set_autodefend_period(unsigned char period);
-unsigned char get_autodefend_period(void);
-eat_bool get_autodefend_state(void);
-
-
 
 #endif /* USER_ELECTROMBILE_DATA_H_ */
